prl-dow30/requirements.txt
numpy>=1.24
pandas>=2.0
yfinance>=0.2
torch>=2.1
stable-baselines3>=2.3.0
gymnasium>=0.29.1
PyYAML>=6.0
matplotlib>=3.8
